+++ database/migrations/2026_07_08_133426_create_katalog_lives_table.php
@@ -8,16 +8,18 @@ return new class extends Migration
 {
     public function up()
     {
-        Schema::create('katalog_live', function (Blueprint $table) {
-            $table->id();
-            $table->date('tanggal_live'); // Sesuai kata Nizar: Relasi berdasarkan tanggal
-            $table->string('kode_live');  // Contoh: "102", "YB 39"
-            $table->unsignedBigInteger('id_produk'); // Relasi ke tabel produk asli
-            $table->integer('stok_alokasi')->default(0); // Contoh: (7), (60)
-            $table->timestamps();
+        if (!Schema::hasTable('katalog_live')) {
+            Schema::create('katalog_live', function (Blueprint $table) {
+                $table->id();
+                $table->date('tanggal_live'); // Sesuai kata Nizar: Relasi berdasarkan tanggal
+                $table->string('kode_live');  // Contoh: "102", "YB 39"
+                $table->unsignedBigInteger('id_produk'); // Relasi ke tabel produk asli
+                $table->integer('stok_alokasi')->default(0); // Contoh: (7), (60)
+                $table->timestamps();
 
-            $table->foreign('id_produk')->references('id')->on('produk')->onDelete('cascade');
-        });
+                $table->foreign('id_produk')->references('id')->on('produk')->onDelete('cascade');
+            });
+        }
     }
 
     public function down()
